requirements.txt
--index-url https://test.pypi.org/simple
--extra-index-url https://pypi.org/simple

intellodge_core==1.0.3
asgiref==3.10.0
awsebcli==3.25.3
bcrypt==5.0.0
blessed==1.24.0
boto==2.49.0
boto3==1.40.52
botocore==1.40.52
cement==2.10.14
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
Django==4.2.25
django-dynamodb-sessions==0.6
django-widget-tweaks==1.5.0
ecdsa==0.19.1
fabric==3.2.2
gunicorn==23.0.0
idna==3.11
invoke==2.2.1
jmespath==1.0.1
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
PyNaCl==1.6.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.15.0
urllib3==1.26.20
wcwidth==0.2.14
wrapt==2.0.1